backend/requirements.txt
fastapi>=0.70.0
uvicorn[standard]>=0.15.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
alembic>=1.7.0
pydantic>=1.8.0
python-multipart>=0.0.5
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pymarc>=4.1.0
qrcode>=7.3
python-barcode>=0.13.1